Refactored IR video and depth video into separate datainterfaces with common parent BaseVideoInterface
