+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023/__init__.py
@@ -3,4 +3,5 @@ from .behavioralsyllableinterface import BehavioralSyllableInterface
 from .fiberphotometryinterface import FiberPhotometryInterface
 from .optogeneticinterface import OptogeneticInterface
 from .depthvideointerface import DepthVideoInterface
+from .irvideointerface import IRVideoInterface
 from .nwbconverter import NWBConverter
